saffron::commitment

Function commit_sparse

pub fn commit_sparse<G: KimchiCurve>(
    srs: &SRS<G>,
    sparse_data: &[G::ScalarField],
    indexes: &[u64],
) -> G

Compute the commitment to the polynomial P of same degree as srs such that P(ω_i) = sparse_data[i] for all ì in indexes and P(ω_i) = 0 for any other ω_i of the domain of the same size as srs. This commitment is computed in a sparse way through the sum of sparse_data[i] × [L_i] for all i in indexes and [L_i] the commitment to the i-th Lagrange polynomial of same degree as srs.